Move make_robot_config
